+++ astarte-dashboard/src/components/Icon.tsx
@@ -42,6 +42,7 @@ const iconToClassName = {
   statusDisconnected: 'fas fa-circle color-red',
   statusKO: 'fas fa-times-circle color-red',
   statusNeverConnected: 'fas fa-circle color-grey',
+  statusInDeletion: 'fas fa-circle color-grey',
   triggers: 'fas fa-bolt',
   policy: 'fas fa-file-invoice',
 };